Avoid volume unexpected deleted from DSW by DSWP when kubelet start
When kubelet starts, pods will be evicted and it may take more time to
reconcile the status in ASW and DSW. When volume is added to ASW and
marked as attached, it may quickly delete by another goroutine DSWP
which leads to volume disapear from ASW and DSW.

Add a requirement here to ask Attachable volume plugin has to be marked
as globally mounted to ensure it won't been deleted by DSWP.

+++ pkg/kubelet/volumemanager/populator/desired_state_of_world_populator.go
@@ -265,6 +265,12 @@ func (dswp *desiredStateOfWorldPopulator) findAndRemoveDeletedPods() {
 					format.Pod(volumeToMount.Pod)), ""))
 			continue
 		}
+
+		if !dswp.actualStateOfWorld.IsVolumeBindMounted(volumeToMount.VolumeName,
+			volumetypes.UniquePodName(volumeToMount.Pod.UID)) && podExists {
+			glog.V(4).Infof(volumeToMount.GenerateMsgDetailed("Actual state has not yet has the volume mount information skip removing volume from desired state", ""))
+			continue
+		}
 		klog.V(4).Infof(volumeToMount.GenerateMsgDetailed("Removing volume from desired state", ""))
 
 		dswp.desiredStateOfWorld.DeletePodFromVolume(
